+++ packages/cubejs-backend-shared/src/index.ts
@@ -26,3 +26,4 @@ export * from './decorators';
 export * from './PerfTracker';
 export * from './disposedProxy';
 export * from './logger';
+export * from './pool';